+++ src/containers/Interfaces/NameGenerator.js
